Do more with ShellCheck (#254)
* Fix numerous issues to make ShellCheck pass across `bin/*` and
  `conf/uno.conf`
* Add GitHub Actions and CI script for automated ShellCheck testing
* Add missing checksum for latest Hadoops still supported by Accumulo
* Fix parameter passing for `--no-deps` flag
* Build SNAPSHOT version with correct version of Guava for Hadoop 3.1
  and later
* Make some long if statements into more concise `&&` syntax for
  readability
* Remove some unnecessary quoting
* Rely on return code for pgrep for many statements that were previously
  relying on its output unnecessarily
* Fix zk dir pattern
* Fix return codes and make rm safer
* Fix ZK capitalization
* Overhaul some of the scripts:
  Consolidate some of the scripts into a commands.sh file and use return
  statements instead of exit statements, as appropriate, for more reliable
  exit behavior for functions.

+++ bin/impl/fetch.sh
@@ -15,6 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# shellcheck source=bin/impl/util.sh
 source "$UNO_HOME"/bin/impl/util.sh
 
 function fetch_hadoop() {
@@ -26,32 +27,33 @@ function fetch_zookeeper() {
 }
 
 function fetch_accumulo() {
-  if [[ $1 != "--no-deps" ]]; then
-    fetch_hadoop
-    fetch_zookeeper
-  fi
+  [[ $1 != '--no-deps' ]] && fetch_hadoop && fetch_zookeeper
 
   if [[ -n "$ACCUMULO_REPO" ]]; then
     declare -a maven_args=(-DskipTests -DskipFormat)
-    if [[ "${HADOOP_VERSION}" = 3.* ]]; then
+    if [[ $HADOOP_VERSION =~ 3\..* ]]; then
       maven_args=("${maven_args[@]}" '-Dhadoop.profile=3')
+      if ! [[ $HADOOP_VERSION =~ 3\.0\..* ]]; then
+        # Hadoop 3.1 and later require a newer version of Guava
+        # than what Accumulo builds by default
+        maven_args=("${maven_args[@]}" '-Dguava.version=27.0-jre')
+      fi
     fi
-    rm -f "$DOWNLOADS/$ACCUMULO_TARBALL"
-    pushd .
-    cd "$ACCUMULO_REPO"
-    mvn -V -e clean package "${maven_args[@]}"
+    rm -f "${DOWNLOADS:?}/${ACCUMULO_TARBALL:?}"
+    (cd "$ACCUMULO_REPO" && mvn -V -e clean package "${maven_args[@]}")
     accumulo_built_tarball=$ACCUMULO_REPO/assemble/target/$ACCUMULO_TARBALL
     if [[ ! -f "$accumulo_built_tarball" ]]; then
-      echo
-      echo "The following file does not exist :"
-      echo "    $accumulo_built_tarball"
-      echo "after building from :"
-      echo "    ACCUMULO_REPO=$ACCUMULO_REPO"
-      echo "ensure ACCUMULO_VERSION=$ACCUMULO_VERSION is correct."
-      echo
+      cat <<EOF
+
+The following file does not exist :
+    $accumulo_built_tarball
+after building from :
+    ACCUMULO_REPO=$ACCUMULO_REPO
+ensure ACCUMULO_VERSION=$ACCUMULO_VERSION is correct.
+
+EOF
       exit 1
     fi
-    popd
     cp "$accumulo_built_tarball" "$DOWNLOADS"/
   else
     download_apache "accumulo/$ACCUMULO_VERSION" "$ACCUMULO_TARBALL" "$ACCUMULO_HASH"
@@ -59,14 +61,10 @@ function fetch_accumulo() {
 }
 
 function fetch_fluo() {
-  if [[ $1 != "--no-deps" ]]; then
-    fetch_accumulo
-  fi
+  [[ $1 != '--no-deps' ]] && fetch_accumulo
   if [[ -n "$FLUO_REPO" ]]; then
-    rm -f "$DOWNLOADS/$FLUO_TARBALL"
-    cd "$FLUO_REPO"
-    mvn -V -e clean package -DskipTests -Dformatter.skip
-
+    rm -f "${DOWNLOADS:?}/${FLUO_TARBALL:?}"
+    (cd "$FLUO_REPO" && mvn -V -e clean package -DskipTests -Dformatter.skip)
     fluo_built_tarball=$FLUO_REPO/modules/distribution/target/$FLUO_TARBALL
     if [[ ! -f "$fluo_built_tarball" ]]; then
       echo "The tarball $fluo_built_tarball does not exist after building from the FLUO_REPO=$FLUO_REPO"
@@ -83,7 +81,7 @@ function fetch_fluo() {
 # Determine best apache mirror to use
 apache_mirror=$(curl -sk https://apache.org/mirrors.cgi?as_json | grep preferred | cut -d \" -f 4)
 
-if [ -z "$apache_mirror" ]; then
+if [[ -z $apache_mirror ]]; then
   echo "Failed querying apache.org for best download mirror!"
   echo "Fetch can only verify existing downloads or build Accumulo/Fluo tarballs from a repo."
 fi
@@ -96,14 +94,10 @@ fluo)
   fetch_fluo "$2"
   ;;
 fluo-yarn)
-  if [[ $2 != "--no-deps" ]]; then
-    fetch_fluo
-  fi
-  if [[ -n "$FLUO_YARN_REPO" ]]; then
-    rm -f "$DOWNLOADS/$FLUO_YARN_TARBALL"
-    cd "$FLUO_YARN_REPO"
-    mvn -V -e clean package -DskipTests -Dformatter.skip
-
+  [[ $2 != '--no-deps' ]] && fetch_fluo
+  if [[ -n $FLUO_YARN_REPO ]]; then
+    rm -f "${DOWNLOADS:?}/${FLUO_YARN_TARBALL:?}"
+    (cd "$FLUO_YARN_REPO" && mvn -V -e clean package -DskipTests -Dformatter.skip)
     built_tarball=$FLUO_YARN_REPO/target/$FLUO_YARN_TARBALL
     if [[ ! -f "$built_tarball" ]]; then
       echo "The tarball $built_tarball does not exist after building from the FLUO_YARN_REPO=$FLUO_YARN_REPO"
@@ -122,13 +116,20 @@ zookeeper)
   fetch_zookeeper
   ;;
 *)
-  echo "Usage: uno fetch <component>"
-  echo -e "\nPossible components:\n"
-  echo "    accumulo   Downloads Accumulo, Hadoop & ZooKeeper. Builds Accumulo if repo set in uno.conf"
-  echo "    fluo       Downloads Fluo, Accumulo, Hadoop & ZooKeeper. Builds Fluo or Accumulo if repo set in uno.conf"
-  echo "    hadoop     Downloads Hadoop"
-  echo "    zookeeper  Downloads ZooKeeper"
-  echo "Options:"
-  echo "    --no-deps  Dependencies will be fetched unless this option is specified. Only works for fluo & accumulo components."
+  cat <<EOF
+Usage: uno fetch <component>
+
+Possible components:
+
+    accumulo   Downloads Accumulo, Hadoop & ZooKeeper. Builds Accumulo if repo set in uno.conf
+    fluo       Downloads Fluo, Accumulo, Hadoop & ZooKeeper. Builds Fluo or Accumulo if repo set in uno.conf
+    hadoop     Downloads Hadoop
+    zookeeper  Downloads ZooKeeper
+
+Options:
+    --no-deps  Dependencies will be fetched unless this option is specified. Only works for fluo & accumulo components.
+EOF
   exit 1
 esac
+
+# fetch.sh
